I have re-worked the approach and it is soooo much cleaner.
Instead of trying to construct contiguous segments and then build a greedy
solution on the resulting dataset, I now take a branch and prune approach. This
will be explained further in the README so as not to clutter the commit log.

I added 3 extra test inputs to test various things. I do not yet automatically
check against the values in the solutions directory, though they do contain the
correct solutions (I believe) to the appropriate inputs.

I have modified the code such that Solution.py is the entry point when not
running via pybuilder. RoverImage.py contains code for managing image chunks,
segments (set of chunks), and the workhorse code for the solver. The solving
code could be moved out into Solution.py but this organization seemed good
enough for now.

I have test code but it is pretty badly written overall. I also need to add
more tests and clean it up quite a bit as far as having good code and tests
goes, but it should test basic functionality of RoverImage and RoverConnection
well enough.