Adding the python file that I was working on at the time of the submission via hacker rank. The code included in this version of CounsylSolution.py does *not* match up with what was submitted to hacker rank because this one doesn't work (though I feel like it's very close), whereas the submitted code "works"--fails correctness on test #4 and fails performance on tests #5-9.
I will add appropriate comments to the top of this file which I should've done
on the hackerrank submission.

Also, CounsylSolution.py is code from src/main/python condensed into a single
file so that it could be submitted via hackerrank. Although perhaps not a great
idea, at the beginning of this challenge I was fairly excited to learn a python
build tool and to attempt an approach to developing that emphasized more
testing. Of course, around Tuesday or so I abandoned that and just tried to get
it to work (because I realized that my initial submission was not satisfying
the given constraints).

Anyways, at least 1 more commit to come regarding comments on algorithmic
complexity.